refactor(ethereum_clis): move generic code from TransitionTool to EthereumCLI
This refactors src/evm_transition_tool/ to src/ethereum_clis/. The aim being to allow easier definition of EVM sub-commands (for use in consume, for example) that can take advantage of a new base class EthereumCLI.
