+++ docs/library/ethereum_clis.md
@@ -1,0 +1,3 @@
+# Ethereum CLIs Package
+
+::: ethereum_clis
