--- docs/library/evm_transition_tool.md
@@ -1,3 +1,0 @@
-# EVM Transition Tool Package
-
-::: evm_transition_tool
