+++ docs/navigation.md
@@ -36,7 +36,7 @@
     * [Running Github Actions Locally](dev/test_actions_locally.md)
   * [Changelog](CHANGELOG.md)
   * [Library Reference](library/index.md)
-    * [Miscellaneous CLI Tools](library/cli/index.md)
+    * [EEST CLI Tools](library/cli/index.md)
     * [Ethereum Test Base Types Package](library/ethereum_test_base_types.md)
     * [Ethereum Test Exceptions Package](library/ethereum_test_exceptions.md)
     * [Ethereum Test Fixtures Package](library/ethereum_test_fixtures.md)
@@ -45,5 +45,5 @@
     * [Ethereum Test Tools Package](library/ethereum_test_tools.md)
     * [Ethereum Test Types Package](library/ethereum_test_types.md)
     * [Ethereum Test VM Package](library/ethereum_test_vm.md)
-    * [EVM Transition Tool Package](library/evm_transition_tool.md)
+    * [Ethereum CLIs Package](library/ethereum_clis.md)
     * [Pytest Plugins](library/pytest_plugins/index.md)
