+++ src/ethereum_clis/tests/fixtures/3/alloc.json
