+++ src/ethereum_clis/transition_tool.py
@@ -10,9 +10,7 @@ import tempfile
 import textwrap
 from abc import abstractmethod
 from dataclasses import dataclass, field
-from itertools import groupby
 from pathlib import Path
-from re import Pattern
 from typing import Dict, List, Mapping, Optional, Type
 
 from requests_unixsocket import Session  # type: ignore
@@ -21,29 +19,21 @@ from ethereum_test_fixtures import FixtureFormat, FixtureVerifier
 from ethereum_test_forks import Fork
 from ethereum_test_types import Alloc, Environment, Transaction
 
+from .clis import EthereumCLI
 from .file_utils import dump_files_to_directory, write_json_file
 from .types import TransactionReceipt, TransitionToolInput, TransitionToolOutput
 
 
-class UnknownTransitionTool(Exception):
+class UnknownCLI(Exception):
     """Exception raised if an unknown t8n is encountered"""
 
     pass
 
 
-class TransitionToolNotFoundInPath(Exception):
-    """Exception raised if the specified t8n tool is not found in the path"""
-
-    def __init__(self, message="The transition tool was not found in the path", binary=None):
-        if binary:
-            message = f"{message} ({binary})"
-        super().__init__(message)
-
-
 model_dump_config: Mapping = {"by_alias": True, "exclude_none": True}
 
 
-class TransitionTool(FixtureVerifier):
+class TransitionTool(EthereumCLI, FixtureVerifier):
     """
     Transition tool abstract base class which should be inherited by all transition tool
     implementations.
@@ -53,9 +43,9 @@ class TransitionTool(FixtureVerifier):
 
     registered_tools: List[Type["TransitionTool"]] = []
     default_tool: Optional[Type["TransitionTool"]] = None
-    default_binary: Path
-    detect_binary_pattern: Pattern
-    version_flag: str = "-v"
+    # default_binary: Path
+    # detect_binary_pattern: Pattern
+    # version_flag: str = "-v"
     t8n_subcommand: Optional[str] = None
     statetest_subcommand: Optional[str] = None
     blocktest_subcommand: Optional[str] = None
@@ -65,8 +55,6 @@ class TransitionTool(FixtureVerifier):
     t8n_use_server: bool = False
     server_url: str
     process: Optional[subprocess.Popen] = None
-
-    # Abstract methods that each tool must implement
 
     @abstractmethod
     def __init__(
@@ -78,17 +66,7 @@ class TransitionTool(FixtureVerifier):
         """
         Abstract initialization method that all subclasses must implement.
         """
-        if binary is None:
-            binary = self.default_binary
-        else:
-            # improve behavior of which by resolving the path: ~/relative paths don't work
-            resolved_path = Path(os.path.expanduser(binary)).resolve()
-            if resolved_path.exists():
-                binary = resolved_path
-        binary = shutil.which(binary)  # type: ignore
-        if not binary:
-            raise TransitionToolNotFoundInPath(binary=binary)
-        self.binary = Path(binary)
+        super().__init__(binary=binary)
         self.trace = trace
 
     def __init_subclass__(cls):
@@ -96,95 +74,6 @@ class TransitionTool(FixtureVerifier):
         Registers all subclasses of TransitionTool as possible tools.
         """
         TransitionTool.register_tool(cls)
-
-    @classmethod
-    def register_tool(cls, tool_subclass: Type["TransitionTool"]):
-        """
-        Registers a given subclass as tool option.
-        """
-        cls.registered_tools.append(tool_subclass)
-
-    @classmethod
-    def set_default_tool(cls, tool_subclass: Type["TransitionTool"]):
-        """
-        Registers the default tool subclass.
-        """
-        cls.default_tool = tool_subclass
-
-    @classmethod
-    def from_binary_path(cls, *, binary_path: Optional[Path], **kwargs) -> "TransitionTool":
-        """
-        Instantiates the appropriate TransitionTool subclass derived from the
-        tool's binary path.
-        """
-        assert cls.default_tool is not None, "default transition tool was never set"
-
-        if binary_path is None:
-            return cls.default_tool(binary=binary_path, **kwargs)
-
-        resolved_path = Path(os.path.expanduser(binary_path)).resolve()
-        if resolved_path.exists():
-            binary_path = resolved_path
-        binary = shutil.which(binary_path)  # type: ignore
-
-        if not binary:
-            raise TransitionToolNotFoundInPath(binary=binary)
-
-        binary = Path(binary)
-
-        # Group the tools by version flag, so we only have to call the tool once for all the
-        # classes that share the same version flag
-        for version_flag, subclasses in groupby(
-            cls.registered_tools, key=lambda x: x.version_flag
-        ):
-            try:
-                result = subprocess.run(
-                    [binary, version_flag], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-                )
-                if result.returncode != 0:
-                    raise Exception(f"Non-zero return code: {result.returncode}")
-
-                if result.stderr:
-                    raise Exception(f"Tool wrote to stderr: {result.stderr.decode()}")
-
-                binary_output = ""
-                if result.stdout:
-                    binary_output = result.stdout.decode().strip()
-            except Exception:
-                # If the tool doesn't support the version flag,
-                # we'll get an non-zero exit code.
-                continue
-            for subclass in subclasses:
-                if subclass.detect_binary(binary_output):
-                    return subclass(binary=binary, **kwargs)
-
-        raise UnknownTransitionTool(f"Unknown transition tool binary: {binary_path}")
-
-    @classmethod
-    def detect_binary(cls, binary_output: str) -> bool:
-        """
-        Returns True if the binary matches the tool
-        """
-        assert cls.detect_binary_pattern is not None
-
-        return cls.detect_binary_pattern.match(binary_output) is not None
-
-    def version(self) -> str:
-        """
-        Return name and version of tool used to state transition
-        """
-        if self.cached_version is None:
-            result = subprocess.run(
-                [str(self.binary), self.version_flag],
-                stdout=subprocess.PIPE,
-            )
-
-            if result.returncode != 0:
-                raise Exception("failed to evaluate: " + result.stderr.decode())
-
-            self.cached_version = result.stdout.decode().strip()
-
-        return self.cached_version
 
     @abstractmethod
     def is_fork_supported(self, fork: Fork) -> bool:
