+++ src/pytest_plugins/consume/direct/conftest.py
@@ -12,10 +12,10 @@ from typing import Generator, Optional
 
 import pytest
 
+from ethereum_clis import TransitionTool
 from ethereum_test_base_types import to_json
 from ethereum_test_fixtures.consume import TestCaseIndexFile, TestCaseStream
 from ethereum_test_fixtures.file import Fixtures
-from evm_transition_tool import TransitionTool
 
 
 def pytest_addoption(parser):  # noqa: D103
